+++ src/pages/app.jsx
@@ -7,6 +7,7 @@ import Map from './map.jsx';
 import Forums from './forums.jsx';
 import '../styles/main.scss';
 import Weather from './weather.jsx';
+import { GetOpenWeatherData } from '../components/openWeatherData.jsx';
 
 
 export default function App() {
@@ -19,6 +20,7 @@ export default function App() {
         <Route exact path="/map" element={<Map />} />
         <Route exact path="/forums" element={<Forums />} />
         <Route exact path="/weather" element={<Weather />} />
+        <Route path="/weather/:city" element={<GetOpenWeatherData />} /> // changes url according to city
       </Routes>
     </Router>
   );